[#100] Rename: reponse -> response 로 변경
