+++ src/main/java/com/vt/valuetogether/domain/team/dto/response/TeamDeleteRes.java
@@ -1,6 +1,6 @@
-package com.vt.valuetogether.domain.team.dto.reponse;
+package com.vt.valuetogether.domain.team.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties
-public class TeamEditRes {}
+public class TeamDeleteRes {}
