+++ src/main/java/com/vt/valuetogether/domain/team/dto/response/TeamEditRes.java
@@ -1,6 +1,6 @@
-package com.vt.valuetogether.domain.team.dto.reponse;
+package com.vt.valuetogether.domain.team.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties
-public class TeamDeleteRes {}
+public class TeamEditRes {}
