+++ src/main/java/com/vt/valuetogether/domain/team/dto/response/TeamMemberInviteRes.java
@@ -1,4 +1,4 @@
-package com.vt.valuetogether.domain.team.dto.reponse;
+package com.vt.valuetogether.domain.team.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
