+++ src/main/java/com/vt/valuetogether/domain/team/service/impl/TeamServiceImpl.java
@@ -1,11 +1,11 @@
 package com.vt.valuetogether.domain.team.service.impl;
 
-import com.vt.valuetogether.domain.team.dto.reponse.TeamCreateRes;
-import com.vt.valuetogether.domain.team.dto.reponse.TeamDeleteRes;
-import com.vt.valuetogether.domain.team.dto.reponse.TeamEditRes;
-import com.vt.valuetogether.domain.team.dto.reponse.TeamGetRes;
-import com.vt.valuetogether.domain.team.dto.reponse.TeamMemberDeleteRes;
-import com.vt.valuetogether.domain.team.dto.reponse.TeamMemberInviteRes;
+import com.vt.valuetogether.domain.team.dto.response.TeamCreateRes;
+import com.vt.valuetogether.domain.team.dto.response.TeamDeleteRes;
+import com.vt.valuetogether.domain.team.dto.response.TeamEditRes;
+import com.vt.valuetogether.domain.team.dto.response.TeamGetRes;
+import com.vt.valuetogether.domain.team.dto.response.TeamMemberDeleteRes;
+import com.vt.valuetogether.domain.team.dto.response.TeamMemberInviteRes;
 import com.vt.valuetogether.domain.team.dto.request.TeamCreateReq;
 import com.vt.valuetogether.domain.team.dto.request.TeamDeleteReq;
 import com.vt.valuetogether.domain.team.dto.request.TeamEditReq;
